policy: Show a simple dialog about optimizing apps
I just really like this alot and cherry-picked it myself, and I dont have any actual data, but it did feel as if apps got optimized a little bit quicker. Try for yourself. :)

NOT MY ORIGINAL CODE OR COMMIT.

http://review.cyanogenmod.org/#/c/131627/

When booting, new apps are optimized with dex2oat. During this process a
progress dialog is displayed (optimizing app 13/74). This dialog has an
indetemninate progress spinner that uses 50% of cpu.

This leaves only 50% for dex2oat, prolonging the time to boot.