Fix navbar secondary overlay resource handling [3/4]
Hold an instance of all our action drawables in an array
in the NavigationController and map actions to needed drawable
index. The secret was to hold instances as persistent members
and only update the drawables when overlay changes. The navbar
overlay is transient and each update is a "one-shot" situation
Use separate public methods for notifying navbar of navbar overlay
change and statusbar recreate change.

Change-Id: I597a33bc8c4c1212bf68e753ce07e2553463b398
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>