OMS7-N: installd: add command 'rmidmap' [8/14]
Add an installd command to remove an idmap file. This is the inverse of
the 'idmap' command and is intended for clean-up once an idmap file is
no longer needed because an APK was removed, etc.

This commit depends on a corresponding commit in frameworks/native (with
the same Change-Id).

Change-Id: I58f55f643da99c0bd69136ee43c1c8c70c352797
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>