[1/5] frameworks/base: Notification Lights
taken from cm: 53ab0dc9a8c195aa3b51d8ae19fbcd0f84537f89

Change-Id: Id8798ced0d39f0826b34d9becb12aaf7e47dbcdd

Lights with Screen On [2/2]: Optional allowment of lights
Implement a setting allowing lights to be activated
for new notifications even if the screen is on.

Lights with screen on and Custom values are separated
in an advanced section for a cleaner overview.

This setting gives the user an oportunity to activate lights
with the screen on and also during DayDream screensaver.
The option is not activated by default.

also:

framework: add config_ledCanPulse overlay setting
Some devices do not support pulsing LEDs. Allow them to hide pulse
settings.

by @romanbb

Change-Id: I812812e94c5c6c4b9614e06c1cb7005ccfc2a625
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>