OMS7-N: Do not expose unused resources in overlay packages [9/14]
Teach ResTable::getResourceIdentifer to ignore overlay packages. When a
client asks for a resource identifier in a package, the lookup should be
made only in the context of that package, i.e. the package and any
related split packages, but excluding any overlay packages.

This prevents the following scenario:

  - target package:
    - drawable/a.png

  - overlay package:
    - drawable/a.png
    - drawable/b.png

ResTable::getResourceIdentifer would return a resource ID even for
drawable/b. The resource ID makes no sense in the context of the target
package, and the actual numerical value is only valid within the context
of the overlay package.

Change-Id: Ic20452b9eac7b1898f2fdc5bfee80a72f0598424
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>